test(services): cover agentic workflow creation
