+++ libs/domains/services/feature/src/index.ts
@@ -111,7 +111,7 @@ export * from './lib/service-blueprint-update-flow/blueprint-update-flow'
 export * from './lib/service-creation-flow/agentic-workflow/agentic-workflow-context'
 export * from './lib/service-creation-flow/agentic-workflow/agentic-workflow-configuration'
 export * from './lib/service-creation-flow/agentic-workflow/agentic-workflow-summary'
-export * from './lib/service-creation-flow/agentic-workflow/use-create-agentic-workflow'
+export * from './lib/service-creation-flow/agentic-workflow/hooks/use-create-agentic-workflow'
 export * from './lib/application-container-healthchecks/application-container-healthchecks-form/application-container-healthchecks-form'
 export * from './lib/application-container-healthchecks/healthchecks-utils'
 export * from './lib/application-container-healthchecks/step-healthchecks/step-healthchecks'
