+++ libs/domains/services/feature/src/lib/service-creation-flow/agentic-workflow/agentic-workflow-configuration/agentic-workflow-configuration.spec.tsx
@@ -1,0 +1,35 @@
+import { getJsonError, isGitRepositoryComplete, isOutputComplete } from './agentic-workflow-configuration'
+
+describe('AgenticWorkflowConfiguration validation', () => {
+  it('should require valid JSON only when a required JSON field is empty or invalid', () => {
+    expect(getJsonError('', false)).toBeUndefined()
+    expect(getJsonError('', true)).toBe('Please enter a valid JSON configuration.')
+    expect(getJsonError('{"mcpServers":{}}', true)).toBeUndefined()
+    expect(getJsonError('{invalid', true)).toBe('Invalid JSON format.')
+  })
+
+  it('should require token, repository, branch, and root path for configured repositories', () => {
+    expect(
+      isGitRepositoryComplete({
+        gitTokenId: 'token-1',
+        repository: 'https://github.com/qovery/console',
+        branch: 'main',
+        rootPath: '/',
+      })
+    ).toBe(true)
+
+    expect(
+      isGitRepositoryComplete({
+        gitTokenId: 'token-1',
+        repository: 'https://github.com/qovery/console',
+        branch: '',
+        rootPath: '/',
+      })
+    ).toBe(false)
+  })
+
+  it('should require a webhook URL for configured output webhooks', () => {
+    expect(isOutputComplete({ url: 'https://hooks.example.com/workflow', headersJson: '{}', prompt: '' })).toBe(true)
+    expect(isOutputComplete({ url: '', headersJson: '{}', prompt: 'Notify the team.' })).toBe(false)
+  })
+})
