+++ libs/domains/services/feature/src/lib/service-creation-flow/agentic-workflow/agentic-workflow-summary.tsx
@@ -8,7 +8,7 @@ import {
   type AgenticWorkflowFormData,
   useAgenticWorkflowCreateContext,
 } from './agentic-workflow-context'
-import { formatAgenticWorkflowRequest, useCreateAgenticWorkflow } from './use-create-agentic-workflow'
+import { formatAgenticWorkflowRequest, useCreateAgenticWorkflow } from './hooks/use-create-agentic-workflow'
 
 function truncateSummary(value: string) {
   if (!value.trim()) return '-'
