+++ libs/domains/services/feature/src/lib/service-creation-flow/agentic-workflow/hooks/use-create-agentic-workflow/use-create-agentic-workflow.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { type AgenticWorkflowHeader, AgenticWorkflowModel, type AgenticWorkflowRequest } from 'qovery-typescript-axios'
 import { mutations } from '@qovery/domains/services/data-access'
 import { queries } from '@qovery/state/util-queries'
-import { type AgenticWorkflowFormData } from './agentic-workflow-context'
+import { type AgenticWorkflowFormData } from '../../agentic-workflow-context'
 
 function formatWhitelistHosts(value: string) {
   return value
